- notification factory for foreground and workers
